Apply changes for refactoring and remove the code while refactoring
