+++ magentacloud25/custom/apps/files_sharing/src/mixins/SharesMixin.js
@@ -31,9 +31,9 @@ import PQueue from 'p-queue'
 import debounce from 'debounce'
 
 import Share from '../../../../../../../nextcloud/apps/files_sharing/src/models/Share.js'
-import SharesRequests from '../../../../../../../nextcloud/apps/files_sharing/src/mixins/ShareRequests.js'
+import SharesRequests from '../mixins/ShareRequests.js'
 import ShareTypes from '../../../../../../../nextcloud/apps/files_sharing/src/mixins/ShareTypes.js'
-import Config from '../services/ConfigService.js'
+import Config from '../services/ConfigServiceCustom.js'
 
 export default {
 	mixins: [SharesRequests, ShareTypes],
