+++ magentacloud25/custom/apps/files_sharing/src/services/ConfigServiceCustom.js
